make schema more readable
